auto-blocks-buyer/requirements.txt
web3>=6.5.0
python-telegram-bot>=21.0
python-dotenv>=1.0.0
requests>=2.31.0


